+++ src/DocumentView.hpp
@@ -342,7 +342,6 @@ private:
     QGraphicsPathItem *ensureSearchItemForPage(int pageno) noexcept;
     QGraphicsPathItem *m_current_search_hit_item{nullptr};
     void updateSelectionPath(int pageno, std::vector<QPolygonF> quads) noexcept;
-    void centerOnPage(int pageno) noexcept;
     QSizeF currentPageSceneSize() const noexcept;
     std::vector<Annotation *> annotationsInArea(int pageno,
                                                 const QRectF &area) noexcept;