+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix fit width not working properly in some cases
 - Fix startup Page widget not respecting fit mode from config
 - Fix scrolling to specific position and center it in the viewport
+- Fix zoom changing the viewport position unexpectedly
 
 
 ## 0.5.2